Rocket spins and "accelerates" (doesn't move but does have cool thrusters), improved a lot of the particle effects stuff.

+++ engine-stuff/Quadrangle.cpp
@@ -24,7 +24,7 @@ void Quadrangle::draw() {
     glBegin(GL_POLYGON);
     glColor3f(getColor().r, getColor().g, getColor().b);
 
-    float theta = getCenter().rotationAngle * M_PI / 180.;
+    double theta = getCenter().rotationAngle * M_PI / 180.;
 
     std::vector<int> xCoords = {-(side1 / 2), (side1 / 2), (side1 / 2), -(side1 / 2)};
